// src/types/offer.ts
import { City } from './city';
import { Host } from './host';
import { Location } from './location';

export type Offer = Omit<OfferPreview, 'previewImage'> &  {
  description: string;
  bedrooms: number;
  goods: string[];
  host: Host;
  images: string[];
  maxAdults: number;
};

export type OfferPreview = {
  id: string;
  title: string;
  type: string;
  price: number;
  city: City;
  location: Location;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage: string;
};